Rejigger CLI database init/upgrades
Move most of the database loading and initialization in the CLI tool
until after arg parsing so that if we are doing something like `--help`
we don't get the lag of running db migration checks.

This also adds --upgrade and --check-upgrades arguments to the CLI that
upgrade and check for upgrades, respectively; migrations are now never
automatically performed, and any other option fails if an upgrade is
required rather than silently upgrading.

This carries through the database code by adding a `check_only` option
to migration code to raise an exception rather than performing upgrades
when an upgrade is needed.

Also includes a DRYing out of recreating the user_permissions view which
was being done in both seqno_etc and room_accessible migrations; moved
it to a separate migration that runs later to recreate.

+++ sogs/migrations/message_views.py
@@ -1,7 +1,8 @@
 import logging
+from .exc import DatabaseUpgradeRequired
 
 
-def migrate(conn):
+def migrate(conn, *, check_only):
     from .. import db
 
     if 'message_metadata' in db.metadata.tables and all(
@@ -11,6 +12,8 @@ def migrate(conn):
         return False
 
     logging.warning("DB migration: recreating message_metadata/message_details views")
+    if check_only:
+        raise DatabaseUpgradeRequired("message views need to be recreated")
 
     conn.execute("DROP VIEW IF EXISTS message_metadata")
     conn.execute("DROP VIEW IF EXISTS message_details")
